change bugs

+++ models/Doctor.py
@@ -7,6 +7,6 @@ class Doctor(db.Model):
     id = db.Column(db.SmallInteger, primary_key=True)
     org = db.Column(db.String)
     fio = db.Column(db.String(200))
-    snils = db.Column(db.Integer, nullable=True)
+    snils = db.Column(db.BigInteger, nullable=True)
     telefon = db.Column(db.String(20), nullable=True)
     spec = db.Column(db.String, nullable=True)
